+++ jwt/common.go
@@ -1,4 +1,4 @@
-package middleware
+package jwt
 
 // middlewareKey is the private type used for goa middlewares to store values in the context.
 // It is private to avoid possible collisions with keys used by other packages.
