+++ jwt/token.go
@@ -1,4 +1,4 @@
-package middleware
+package jwt
 
 import (
 	"time"
